+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- @alexjercan implement OnEnter and OnExit events
+
 ## [0.2.1] - 2025-11-15
 
 ### Chores